video: msm: Fix MDP bus bandwidth calculation
 * Set values for 8064 (JF) in board file
 * Remove redundant calculations
 * Use 64-bit unsigned long values where appropriate

+++ arch/arm/mach-msm/include/mach/board.h
@@ -403,7 +403,7 @@ struct msm_panel_common_pdata {
 	int (*vga_switch)(int select_vga);
 	int *gpio_num;
 	u32 mdp_max_clk;
-	u32 mdp_max_bw;
+	u64 mdp_max_bw;
 	u32 mdp_bw_ab_factor;
 	u32 mdp_bw_ib_factor;
 #ifdef CONFIG_MSM_BUS_SCALING
